Move SystemPreparationConfig to configuration module
