+++ a3fe/configuration/__init__.py
@@ -1,0 +1,3 @@
+"""Pydantic configuration classes for the a3fe package."""
+
+from .system_preparation import SystemPreparationConfig
